Add files via upload
An analysis on hypertension and cholesterol values that were measured and collected by the World Health Organization. The notebook includes the links to the datasets, data cleaning, as well as converting the datasets to SQL tables in order to extract the necessary data for analysis. Discussion paragraphs are also included to explain my thought process as well the the approach to the coding. The data is also specific to Canada as this was my group's home country and we believed it was important to investigate how Canada's health compares to the rest of the world. This analysis was a part of a group project in which my responsibilities included answering the following questions "How do cholesterol levels influence the prevalence of hypertension? How do these factors correlate to the numbers of deaths attributed to cardiovascular disease?".